app/Python/requirements.txt
flask
mashmallow
sqlalchemy
mysql.connector
mysql-connector-python
